requirements.txt
flask
flask_limiter
confluent_kafka
pyspark
elasticsearch
python-dotenv